expose this library as global.IntlPolyfill
The rational behind this is to give the user control over which version
of the Intl library to use. The environment (node, browser, etc) might already
have an `Intl` global, and the user might want to use some of its APIs
and some from this library.  The point is we don't know the needs of the user
so lets not presume to much.  Let the user decide if/when/how this libarary
will be used.

To help support the user we should also add a quick/easy "getting started"
section in README.md.  This would be good regardless of whether this pull
request lands, so I'll submit a separate PR for that :)