Rewrite createRegExpRestore for shorter expressions
Previously createRegExpRestore build a regex with the exact characters for each part of the expression (either capturing group, or segment between capturing groups).  The new solution begins with this approach, and then shortens the resulting expression by replacing each segment with a match for character count only, then setting lastIndex on the resultant RegExp to ensure that the match occurs at the expected location in the input.